requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
cattrs==25.1.1
cffi==1.17.1
discord.py==2.6.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.41
idna==3.10
multidict==6.6.4
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.1
setuptools==75.6.0
smmap==5.0.2
typing_extensions==4.14.1
yarl==1.20.1
